requirements.txt
alembic==1.9.3
argcomplete==2.0.0
asgiref==3.7.2
attrs==22.2.0
Authlib==1.2.1
azure-common==1.1.28
azure-core==1.28.0
azure-identity==1.10.0
azure-mgmt-core==1.3.2
azure-mgmt-rdbms==10.2.0b10
azure-mgmt-resource==23.1.0b2
azure-mgmt-subscription==3.0.0
Babel==2.11.0
bcrypt==3.2.2
Beaker==1.12.1
beautifulsoup4==4.12.2
bidict==0.22.0
blinker==1.5
blivet==3.7.1
blivet-gui==2.4.1
boto3==1.26.111
botocore==1.29.111
Brlapi==0.8.4
Brotli==1.0.9
cached-property==1.5.2
certifi==2022.9.24
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
cryptography==37.0.2
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dnspython==2.3.0
docker==5.0.3
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
email-validator==1.1.3
eventlet==0.33.3
fedora-third-party==0.10
file-magic==0.4.0
flake8==6.0.0
flake8-isort==6.0.0
Flask==2.2.3
Flask-Babel==2.0.0
flask-compress==1.13
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.5
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
fros==1.1
gpg==1.17.1
greenlet==1.1.3
gssapi==1.7.3
httpagentparser==1.9.5
humanize==3.13.1
idna==3.4
isodate==0.6.1
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.0.3
jmespath==1.0.1
jsonschema==4.17.3
langtable==0.0.62
ldap3==2.9.1
libcomps==0.1.18
lxml==4.9.2
Mako==1.2.3
MarkupSafe==2.1.2
mccabe==0.7.0
msal==1.22.0
msal-extensions==1.0.0
msrest==0.7.1
nftables==0.1
oauthlib==3.2.2
olefile==0.46
packaging==23.0
paramiko==3.1.0
passlib==1.7.4
Paste==3.5.2
pexpect==4.8.0
pid==2.2.3
Pillow==9.4.0
ply==3.11
portalocker==2.7.0
productmd==1.36
psutil==5.9.5
psycopg2==2.9.6
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pwquality==1.4.5
pyasn1==0.4.8
pycairo==1.23.0
pycodestyle==2.10.0
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.2
pyenchant==3.2.2
pyflakes==3.0.1
PyGObject==3.44.1
PyJWT==2.8.0
pykickstart==3.47
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyotp==2.9.0
pyparted==3.12.0
pyrsistent==0.19.3
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-engineio==4.5.1
python-meh==0.51
python-socketio==5.8.0
python3-openid==3.2.0
pytz==2023.3
pyudev==0.24.0
pyxdg==0.27
PyYAML==6.0
qrcode==7.3.1
regex==2023.6.3
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
requests-oauthlib==1.3.1
rpm==4.18.1
s3transfer==0.6.1
selinux==3.5
sepolicy==3.5
setools==4.4.2
simpleaudio==1.0.4
simplejson==3.18.3
simpleline==1.9.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sos==4.5.1
soupsieve==2.4.1
SQLAlchemy==1.4.49
sqlparse==0.4.4
sshtunnel==0.4.0
systemd-python==235
Tempita==0.5.2
texttable==1.6.4
typing_extensions==4.5.0
ua-parser==0.18.0
ufw==0.35
urllib3==1.26.16
user-agents==2.2.0
websocket-client==1.3.3
Werkzeug==2.2.3
WTForms==3.0.1
